+++ packages/stream_video/lib/stream_video.dart
@@ -51,6 +51,7 @@ export 'src/webrtc/media/media_constraints.dart';
 export 'src/webrtc/model/rtc_video_dimension.dart';
 export 'src/webrtc/model/rtc_video_parameters.dart';
 export 'src/webrtc/model/stats/rtc_stats_models.dart';
+export 'src/webrtc/model/track_disable_mode.dart';
 export 'src/webrtc/peer_type.dart';
 export 'src/webrtc/rtc_media_device/rtc_media_device.dart';
 export 'src/webrtc/rtc_media_device/rtc_media_device_notifier.dart';